Address comments and fix tests
