+++ tests/integration/test_slide.py
@@ -7,10 +7,12 @@ import numpy as np
 import PIL
 import pytest
 
+from histolab.exceptions import LevelError, SlidePropertyError
 from histolab.masks import BiggestTissueBoxMask, TissueMask
 from histolab.slide import Slide
 
-from ..fixtures import SVS
+from ..fixtures import EXTERNAL_SVS, SVS
+from ..unitutil import on_ci
 from ..util import load_expectation, load_python_expression
 
 
@@ -47,6 +49,80 @@ class Describe_Slide:
         assert image.size == dimensions
         assert slide.dimensions == (2220, 2967)
         assert image.size == (2220, 2967)
+
+    @pytest.mark.parametrize(
+        "fixture_slide, level, expected_value",
+        (
+            pytest.param(
+                EXTERNAL_SVS.LIVER_1,
+                0,
+                "20.0X",
+                marks=pytest.mark.skipif(not on_ci(), reason="To run only on CI"),
+            ),
+            pytest.param(
+                EXTERNAL_SVS.LIVER_1,
+                1,
+                "5.0X",
+                marks=pytest.mark.skipif(not on_ci(), reason="To run only on CI"),
+            ),
+            pytest.param(
+                EXTERNAL_SVS.LIVER_1,
+                2,
+                "1.25X",
+                marks=pytest.mark.skipif(not on_ci(), reason="To run only on CI"),
+            ),
+        ),
+    )
+    def it_knows_its_magnification_factors(self, fixture_slide, level, expected_value):
+        slide = Slide(fixture_slide, "")
+
+        magnification_factor = slide.level_magnification_factor(level=level)
+
+        assert magnification_factor == expected_value
+
+    @pytest.mark.parametrize(
+        "fixture_slide, level",
+        (
+            (SVS.TCGA_CR_7395_01A_01_TS1, 0),
+            (SVS.TCGA_CR_7395_01A_01_TS1, 1),
+        ),
+    )
+    def but_it_raises_exception_when_magnification_cannot_be_computed(
+        self, fixture_slide, level
+    ):
+        slide = Slide(fixture_slide, "")
+
+        with pytest.raises(SlidePropertyError) as err:
+            slide.level_magnification_factor(level=level)
+
+        assert isinstance(err.value, SlidePropertyError)
+        assert (
+            str(err.value)
+            == f"Native magnification not available. Available slide properties: "
+            f"{list(slide.properties.keys())}"
+        )
+
+    @pytest.mark.parametrize(
+        "fixture_slide, level",
+        [
+            pytest.param(
+                EXTERNAL_SVS.LIVER_1,
+                4,
+                marks=pytest.mark.skipif(not on_ci(), reason="To run only on CI"),
+            ),
+        ],
+    )
+    def and_when_level_is_incorrect(self, fixture_slide, level):
+        slide = Slide(fixture_slide, "")
+
+        with pytest.raises(LevelError) as err:
+            slide.level_magnification_factor(level=level)
+
+        assert isinstance(err.value, LevelError)
+        assert (
+            str(err.value)
+            == f"Level {level} not available. Number of available levels: 3"
+        )
 
     def it_raises_openslideerror_with_broken_wsi(self):
         slide = Slide(SVS.BROKEN, os.path.join(SVS.BROKEN, "processed"))
